pyproject.toml
[project]
name = "aws-agent-wordpress"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.2.11",
    "uvicorn>=0.35.0",
]
